improved OpenMP usage
1) reduced the number of fork-join per iteration

'omp parallel for' does a fork-join, which can get expensive at large
thread-counts.  when this construct is used many times in a function, it
should be replaced with a single 'omp parallel' around multiple 'omp
for'.

the code previously found between parallel regions is assumed to
require serialization and uses 'pragma omp single' for protectin.
'single' is used instead of 'master' to allow the first encountering
thread in the team to do the work, rather than waiting for the master
thread.

technically, but never in practice, 'single' requires
MPI_THREAD_SERIALZIED instead of MPI_THREAD_FUNNELED.  'master' only
requires MPI_THREAD_FUNNELED.

it is possible that 'single nowait' is sufficient, in which case a few
barriers can be eliminated.  (aside: 'master' does not imply a barrier).

2) pragma omp simd wherever pragma ivdep is used

the OpenMP standard defines 'pragma omp simd' semantics identical to the
convention meaning of the non-standard 'pragma ivdep'.

Intel compiler treats 'pragma omp simd' as an assertion rather than a
hint so if SIMD isn't appropriate, this pragma should be conditionalized
using preprocessor (C99/C++11 _Pragma being the O(1) solution here).