Implement jit-ed version of gpuMain2 (the dominant kernel) and all its supporting device functions.
All accesses of __constant__ memory have been replaced by jit-ed values.

Jit-ed kernel is not invoked yet, as there is one thing left to do:
deal with "numsbad", a variable in __constant__ memory that is
modified by the kernel (!) and read back to the host afterward. So we
can't jit this variable. Instead, we'll move it to regular device memory.