Start implementation of zone-based version of gpuMain2.
Implementation strategy:
Gradually implement gpuMain2_zb. For any data structure 'data' that the original
gpuMain2 modifies, we'll create a mirror version 'data_zb, to be
used/modified by the new kernel. Mirroring takes place in every iteration of
the hydro cycle loop, prior to calling both gpuMain2 and gpuMain2_zb. We also
create an additional mirror 'data_cp', which is checkpoint data, to be written
by the original gpuMain2. We do this since 'data' may be modified multiple times
by gpuMain2, and we want to keep the state that correcponds to the (possibly
incomplete) state of the corresponding 'data_zb', as computed by the partially
implemented gpuMain2_zb.

After every iteration of gpuMain2 and gpuMain_zb, we verify that data_zb is
identical to data_cp. Once the implementation of gpuMain2_zb is correct,
we'll swap it out for the original gpuMain2, let it operate on the original data,
and remove the original gpuMain2 and all the temporary _zb and _cp data structures.