Optimize zone-based version of meshCalcCharLen.
This optimization reuses data from earlier loop iterations in
subsequent iterations, saving some look-ups. The compiler can't
deduce that this data can be reused.

This optimization leads to a 10% savings of cycles for the partially
implemented zone-based version of gpuMain2, but it turns out to be
hardly noticeable in the overal hydro cycle time. Also, numerical
results are marginally different tan the unoptimzed version, probably
due to computation reordering. For this reason, I'll undo the
optimzation in the next commit, to make correctness validation for the
complete zone-based gpuMain2 a bit easier. Once gpuMain2 has been
completely rewritten, I'll revisit this optimization.