Add __launch_bounds__ to all kernels.
Launch bounds are needed to prevent a performance drop in ROCm versions after
the compiler defaults from max 256 threads/block to max 1024, which increases
register pressure.
One of the kernels conditionally has a second launch bound parameter for GFX908 (MI100).
This forces spilling to the AccVGPR register file, which is unique to MI100.