Replace hipEventSynchronize by synchronization using coherence and atomics.
When GPU is done computing the delta-t value for the next iteration,
it writes it to host-pinned memory, and raises an atomic flag in
host-pinned memory. On the host side, the (heavy) hipEventSynchronize
(or alternatively, hipDeviceSynchronize), is replace by spin-looping
on the flag that is raised by the GPU. Once the flag is up, we can safely
read the data written by the GPU into host-pinned memory.