Revert an earlier change that introduced a bug.
The earlier change lifted the computation of 'fac' out of the GPU code. This
was only half of the intended code change. The other half was not integrated
properly, and changed the value of znump[z] to 4, except if the value was
3, in which case it was left unchanged. While these two changes together gave
correct results, znump now no longer correctly represented the number of
points in a zone. Any code that would rely on this variable meaning the number
of points in a zone would now be incorrect. There is currently no such use in
HydroGPU.cc, but any future use would lead to hard to find bugs.

The proper way to lift this computation out of the GPU code is to introduce a
new member in the Mesh struct specifically for this purpose. Since that is a
bit more work, with little to no performance benefits, we defer that change.