Flip some CMake variable from advanced to non-advanced and vice versa.
HIP introduced a new variable that we don't necessarily care about.
Showing some MPI variables to be able to verify that we are building with
the expected MPI version, which becomes especially important when trying
ROCm-aware MPI, as the code and the MPI libraries should agree about being
able to use device pointers for MPI communications.